requirements.txt
streamlit
pillow
pandas
openai
transformers
torch
scikit-learn
requests
